requirements.txt
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
pyTelegramBotAPI==4.12.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
django-filter==23.2